save state

+++ WeatherApp/data/network/weather/pojo/response/MainItem.cs
@@ -4,8 +4,8 @@ namespace WeatherApp.data.network.dto.response
     public class MainItem
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
+        public double pressure { get; set; }
+        public double humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
     }
